requirements.txt
pillow
requests
python-dotenv
openai